+++ packages/wix-ui-core/src/components/Button/index.ts
@@ -1,1 +1,2 @@
-export {Button, ButtonProps} from './Button';
+// Backward compatibility for deprecated components
+export * from '../../components/deprecated/Button';
