+++ packages/wix-ui-core/src/components/deprecated/Button/index.ts
@@ -1,0 +1,1 @@
+export {Button, ButtonProps} from './Button';
