+++ packages/wix-ui-core/src/testkit/enzyme.ts
@@ -62,7 +62,7 @@ export const toggleSwitchTestkitFactory = enzymeTestkitFactoryCreator(
 import {
   buttonDriverFactory,
   ButtonDriver
-} from "../components/Button/Button.driver";
+} from "../components/deprecated/Button/Button.driver";
 export const buttonTestkitFactory = enzymeTestkitFactoryCreator<ButtonDriver>(
   buttonDriverFactory
 );
